+++ packages/@jsii/kernel/src/kernel.test.ts
@@ -16,11 +16,9 @@ import {
   WireStruct,
   TOKEN_STRUCT,
 } from './api';
-import { DiskCache } from './disk-cache';
 import { Kernel } from './kernel';
 import { closeRecording, recordInteraction } from './recording';
 import * as tar from './tar-cache';
-import { defaultCacheRoot } from './tar-cache/default-cache-root';
 
 /* eslint-disable require-atomic-updates */
 
@@ -53,15 +51,6 @@ if (recordingOutput) {
   fs.mkdirpSync(recordingOutput);
   console.error(`JSII_RECORD=${recordingOutput}`);
 }
-
-afterAll(() => {
-  // Jest prevents execution of "beforeExit" events. Honour the per-test-run
-  // cache root override (set by jest's globalSetup) so we don't touch the
-  // developer's real package cache.
-  DiskCache.inDirectory(
-    process.env.JSII_RUNTIME_PACKAGE_CACHE_ROOT ?? defaultCacheRoot(),
-  ).pruneExpiredEntries();
-});
 
 function defineTest(
   name: string,
