+++ packages/@jsii/runtime/webpack.config.js
@@ -14,11 +14,11 @@ module.exports = {
     },
     // Worker entry that builds the kernel's runtime index off the main thread.
     // Emitted as a sibling of program.js so the kernel can start it via
-    // `new Worker(join(__dirname, 'runtime-index-builder.js'))`. The module
+    // `new Worker(join(__dirname, 'runtime-index-worker.js'))`. The module
     // self-executes on a worker thread (it is guarded by `!isMainThread`).
-    'index-builder': {
-      import: '@jsii/kernel/lib/runtime-index-builder',
-      filename: 'lib/runtime-index-builder.js',
+    'index-worker': {
+      import: '@jsii/kernel/lib/runtime-index-worker',
+      filename: 'lib/runtime-index-worker.js',
     },
   },
   output: {
